Revert "PR for EH + serializer integration (#12205)"
This reverts commit 7d211b4699e17156bca3ccb6a5e11b9b2644b75b.
