+++ sdk/eventhubs/azure-messaging-eventhubs/src/main/java/com/azure/messaging/eventhubs/EventHubPartitionAsyncConsumer.java
@@ -4,7 +4,6 @@
 package com.azure.messaging.eventhubs;
 
 import com.azure.core.amqp.implementation.MessageSerializer;
-import com.azure.core.experimental.serializer.ObjectSerializer;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.messaging.eventhubs.implementation.AmqpReceiveLinkProcessor;
 import com.azure.messaging.eventhubs.models.EventPosition;
@@ -40,16 +39,13 @@ class EventHubPartitionAsyncConsumer implements AutoCloseable {
     private final Scheduler scheduler;
     private final EmitterProcessor<PartitionEvent> emitterProcessor;
     private final EventPosition initialPosition;
-    private final ObjectSerializer serializer;
 
     private volatile Long currentOffset;
 
     EventHubPartitionAsyncConsumer(AmqpReceiveLinkProcessor amqpReceiveLinkProcessor,
-                                       MessageSerializer messageSerializer, String fullyQualifiedNamespace,
-                                       String eventHubName, String consumerGroup, String partitionId,
-                                       ObjectSerializer serializer,
-                                       AtomicReference<Supplier<EventPosition>> currentEventPosition,
-                                       boolean trackLastEnqueuedEventProperties, Scheduler scheduler) {
+        MessageSerializer messageSerializer, String fullyQualifiedNamespace, String eventHubName, String consumerGroup,
+        String partitionId, AtomicReference<Supplier<EventPosition>> currentEventPosition,
+        boolean trackLastEnqueuedEventProperties, Scheduler scheduler) {
         this.initialPosition = Objects.requireNonNull(currentEventPosition.get().get(),
             "'currentEventPosition.get().get()' cannot be null.");
         this.amqpReceiveLinkProcessor = amqpReceiveLinkProcessor;
@@ -58,7 +54,6 @@ class EventHubPartitionAsyncConsumer implements AutoCloseable {
         this.eventHubName = eventHubName;
         this.consumerGroup = consumerGroup;
         this.partitionId = partitionId;
-        this.serializer = serializer;
         this.trackLastEnqueuedEventProperties = trackLastEnqueuedEventProperties;
         this.scheduler = Objects.requireNonNull(scheduler, "'scheduler' cannot be null.");
 
@@ -143,6 +138,6 @@ class EventHubPartitionAsyncConsumer implements AutoCloseable {
 
         final PartitionContext partitionContext = new PartitionContext(fullyQualifiedNamespace, eventHubName,
             consumerGroup, partitionId);
-        return new PartitionEvent(partitionContext, event, lastEnqueuedEventProperties.get(), this.serializer);
+        return new PartitionEvent(partitionContext, event, lastEnqueuedEventProperties.get());
     }
 }
