+++ sdk/eventhubs/azure-messaging-eventhubs/src/main/java/module-info.java
@@ -3,7 +3,6 @@
 
 module com.azure.messaging.eventhubs {
     requires transitive com.azure.core;
-    requires transitive com.azure.core.experimental;
     requires transitive com.azure.core.amqp;
 
     requires com.microsoft.azure.qpid.protonj.extensions;
