+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/EventHubPartitionAsyncConsumerTest.java
@@ -8,7 +8,6 @@ import com.azure.core.amqp.AmqpMessageConstant;
 import com.azure.core.amqp.AmqpRetryPolicy;
 import com.azure.core.amqp.implementation.AmqpReceiveLink;
 import com.azure.core.amqp.implementation.MessageSerializer;
-import com.azure.core.experimental.serializer.ObjectSerializer;
 import com.azure.core.util.Context;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.messaging.eventhubs.implementation.AmqpReceiveLinkProcessor;
@@ -31,12 +30,9 @@ import reactor.core.Disposables;
 import reactor.core.publisher.DirectProcessor;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.FluxSink;
-import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
 import reactor.test.StepVerifier;
 
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Date;
@@ -121,7 +117,7 @@ class EventHubPartitionAsyncConsumerTest {
         // Arrange
         linkProcessor = createSink(link1, link2).subscribeWith(new AmqpReceiveLinkProcessor(PREFETCH, retryPolicy, parentConnection));
         consumer = new EventHubPartitionAsyncConsumer(linkProcessor, messageSerializer, HOSTNAME, EVENT_HUB_NAME,
-            CONSUMER_GROUP, PARTITION_ID, null, currentPosition, trackLastEnqueuedProperties, Schedulers.parallel());
+            CONSUMER_GROUP, PARTITION_ID, currentPosition, trackLastEnqueuedProperties, Schedulers.parallel());
 
         final EventData event1 = new EventData("Foo");
         final EventData event2 = new EventData("Bar");
@@ -169,7 +165,7 @@ class EventHubPartitionAsyncConsumerTest {
         // Arrange
         linkProcessor = createSink(link1, link2).subscribeWith(new AmqpReceiveLinkProcessor(PREFETCH, retryPolicy, parentConnection));
         consumer = new EventHubPartitionAsyncConsumer(linkProcessor, messageSerializer, HOSTNAME, EVENT_HUB_NAME,
-            CONSUMER_GROUP, PARTITION_ID, null, currentPosition, false, Schedulers.parallel());
+            CONSUMER_GROUP, PARTITION_ID, currentPosition, false, Schedulers.parallel());
 
         final Message message3 = mock(Message.class);
         final String secondOffset = "54";
@@ -228,57 +224,6 @@ class EventHubPartitionAsyncConsumerTest {
         Assertions.assertFalse(actual.isInclusive());
     }
 
-    @Test
-    void receiveAndDeserialize() {
-        // just a test value
-        Object o = 0;
-        ObjectSerializer testSerializer = new ObjectSerializer() {
-            @Override
-            public <T> Mono<T> deserialize(InputStream stream, Class<T> clazz) {
-                if (clazz.isInstance(o)) {
-                    return Mono.just(clazz.cast(o));
-                }
-                return null;
-            }
-
-            @Override
-            public <S extends OutputStream> Mono<S> serialize(S stream, Object value) {
-                return null;
-            }
-        };
-
-        // Arrange
-        linkProcessor = createSink(link1, link2).subscribeWith(new AmqpReceiveLinkProcessor(PREFETCH, retryPolicy, parentConnection));
-        consumer = new EventHubPartitionAsyncConsumer(linkProcessor, messageSerializer, HOSTNAME, EVENT_HUB_NAME,
-            CONSUMER_GROUP, PARTITION_ID, testSerializer, currentPosition, false, Schedulers.parallel());
-
-        final EventData event = new EventData("Foo");
-        final LastEnqueuedEventProperties last = new LastEnqueuedEventProperties(10L, 15L,
-            Instant.ofEpochMilli(1243454), Instant.ofEpochMilli(1240004));
-
-        when(messageSerializer.deserialize(same(message1), eq(EventData.class))).thenReturn(event);
-        when(messageSerializer.deserialize(same(message1), eq(LastEnqueuedEventProperties.class))).thenReturn(last);
-
-        // Act & Assert
-        StepVerifier.create(consumer.receive())
-            .then(() -> {
-                messageProcessorSink.next(message1);
-            })
-            .assertNext(partitionEvent -> {
-                verifyPartitionContext(partitionEvent.getPartitionContext());
-                verifyLastEnqueuedInformation(false, last,
-                    partitionEvent.getLastEnqueuedEventProperties());
-                Assertions.assertSame(event, partitionEvent.getData());
-                Assertions.assertSame(Integer.class.cast(o), partitionEvent.getDeserializedObject(Integer.class).block());
-            })
-            .thenCancel()
-            .verify();
-
-        // The emitter processor is not closed until the partition consumer is.
-        Assertions.assertFalse(linkProcessor.isTerminated());
-        Assertions.assertSame(originalPosition, currentPosition.get().get());
-    }
-
 
     /**
      * Verifies that the consumer closes and completes any listeners on a shutdown signal.
@@ -288,7 +233,7 @@ class EventHubPartitionAsyncConsumerTest {
         // Arrange
         linkProcessor = createSink(link1, link2).subscribeWith(new AmqpReceiveLinkProcessor(PREFETCH, retryPolicy, parentConnection));
         consumer = new EventHubPartitionAsyncConsumer(linkProcessor, messageSerializer, HOSTNAME, EVENT_HUB_NAME,
-            CONSUMER_GROUP, PARTITION_ID, null, currentPosition, false, Schedulers.parallel());
+            CONSUMER_GROUP, PARTITION_ID, currentPosition, false, Schedulers.parallel());
 
         final Message message3 = mock(Message.class);
         final String secondOffset = "54";
