+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/EventProcessorClientErrorHandlingTest.java
@@ -139,7 +139,7 @@ public class EventProcessorClientErrorHandlingTest {
 
     private PartitionEvent getEvent(EventData event) {
         PartitionContext context = new PartitionContext("ns", "foo", "bar", "baz");
-        return new PartitionEvent(context, event, null, null);
+        return new PartitionEvent(context, event, null);
     }
