--- sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/ObjectBatchTest.java
@@ -1,54 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.messaging.eventhubs;
-
-import com.azure.core.amqp.implementation.ErrorContextProvider;
-import com.azure.core.experimental.serializer.ObjectSerializer;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import reactor.core.publisher.Mono;
-
-import java.io.InputStream;
-import java.io.OutputStream;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-public class ObjectBatchTest {
-    @Mock
-    private ErrorContextProvider errorContextProvider;
-
-    @BeforeEach
-    public void setup() {
-        MockitoAnnotations.initMocks(this);
-    }
-
-    @Test
-    public void nullObject() {
-        assertThrows(IllegalArgumentException.class, () -> {
-            final ObjectBatch<Object> batch = new ObjectBatch<>(
-                1024,
-                null,
-                null,
-                Object.class,
-                errorContextProvider,
-                null,
-                new ObjectSerializer() {
-                    @Override
-                    public <T> Mono<T> deserialize(InputStream stream, Class<T> clazz) {
-                        return null;
-                    }
-
-                    @Override
-                    public <S extends OutputStream> Mono<S> serialize(S stream, Object value) {
-                        return null;
-                    }
-                },
-                null,
-                null);
-            batch.tryAdd(null);
-        });
-    }
-}
