+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/PartitionBasedLoadBalancerTest.java
@@ -113,7 +113,7 @@ public class PartitionBasedLoadBalancerTest {
             .thenReturn(Flux.interval(Duration.ofSeconds(1)).map(index -> {
                 final PartitionContext partitionContext = new PartitionContext("ns", "foo", "bar", "bazz");
                 int i = index.intValue() % eventDataList.size();
-                return new PartitionEvent(partitionContext, eventDataList.get(i), null, null);
+                return new PartitionEvent(partitionContext, eventDataList.get(i), null);
             }));
 
         PartitionBasedLoadBalancer partitionBasedLoadBalancer = createPartitionLoadBalancer("owner1");
@@ -149,7 +149,7 @@ public class PartitionBasedLoadBalancerTest {
             .thenReturn(Flux.interval(Duration.ofSeconds(1)).map(index -> {
                 final PartitionContext partitionContext = new PartitionContext("ns", "foo", "bar", "bazz");
                 int i = index.intValue() % eventDataList.size();
-                return new PartitionEvent(partitionContext, eventDataList.get(i), null, null);
+                return new PartitionEvent(partitionContext, eventDataList.get(i), null);
             }));
 
         PartitionBasedLoadBalancer partitionBasedLoadBalancer1 = createPartitionLoadBalancer("owner1");
@@ -185,7 +185,7 @@ public class PartitionBasedLoadBalancerTest {
             .thenReturn(Flux.interval(Duration.ofSeconds(1)).map(index -> {
                 final PartitionContext partitionContext = new PartitionContext("ns", "foo", "bar", "bazz");
                 int i = index.intValue() % eventDataList.size();
-                return new PartitionEvent(partitionContext, eventDataList.get(i), null, null);
+                return new PartitionEvent(partitionContext, eventDataList.get(i), null);
             }));
 
         PartitionBasedLoadBalancer partitionBasedLoadBalancer1 = createPartitionLoadBalancer("owner1");
@@ -225,7 +225,7 @@ public class PartitionBasedLoadBalancerTest {
             .thenReturn(Flux.interval(Duration.ofSeconds(1)).map(index -> {
                 final PartitionContext partitionContext = new PartitionContext("ns", "foo", "bar", "bazz");
                 int i = index.intValue() % eventDataList.size();
-                return new PartitionEvent(partitionContext, eventDataList.get(i), null, null);
+                return new PartitionEvent(partitionContext, eventDataList.get(i), null);
             }));
 
         List<PartitionBasedLoadBalancer> loadBalancers = new ArrayList<>();
@@ -263,7 +263,7 @@ public class PartitionBasedLoadBalancerTest {
             .thenReturn(Flux.interval(Duration.ofSeconds(1)).map(index -> {
                 final PartitionContext partitionContext = new PartitionContext("ns", "foo", "bar", "bazz");
                 final int i = index.intValue() % eventDataList.size();
-                return new PartitionEvent(partitionContext, eventDataList.get(i), null, null);
+                return new PartitionEvent(partitionContext, eventDataList.get(i), null);
             }));
 
         List<PartitionBasedLoadBalancer> loadBalancers = new ArrayList<>();
@@ -422,7 +422,7 @@ public class PartitionBasedLoadBalancerTest {
             .thenReturn(Flux.interval(Duration.ofSeconds(1)).map(index -> {
                 final PartitionContext partitionContext = new PartitionContext("ns", "foo", "bar", "bazz");
                 final int i = index.intValue() % eventDataList.size();
-                return new PartitionEvent(partitionContext, eventDataList.get(i), null, null);
+                return new PartitionEvent(partitionContext, eventDataList.get(i), null);
             }));
 
         String ownerName = "owner1";
