+++ sdk/schemaregistry/azure-data-schemaregistry-avro/src/main/java/com/azure/data/schemaregistry/avro/SchemaRegistryAvroAsyncSerializer.java
@@ -3,6 +3,7 @@
 
 package com.azure.data.schemaregistry.avro;
 
+import com.azure.data.schemaregistry.AbstractDataSerializer;
 import com.azure.data.schemaregistry.SerializationException;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Scheduler;
@@ -13,7 +14,7 @@ import java.util.concurrent.Executors;
 /**
  * Asynchronous registry-based serializer implementation.
  */
-public class SchemaRegistryAvroAsyncSerializer {
+public class SchemaRegistryAvroAsyncSerializer extends AbstractDataSerializer {
     private static final int DEFAULT_THREAD_POOL_SIZE = 8;
 
     private final SchemaRegistryAvroSerializer serializer;
